Add new greeting and make it responsive
