+++ client/src/pages/Home/style.scss
@@ -5,9 +5,6 @@ $gold: #D8B646;
 $light-gold: #E2CA78;
 
 .home-sprite-wrapper {
-    width: 15rem;
-    margin-top: 6rem; 
-    padding-bottom: 1.5rem; 
     animation: hover-1 5s ease-in-out infinite;  
 
     svg {
@@ -16,6 +13,9 @@ $light-gold: #E2CA78;
     }
 }
 
+.optional-margin {
+    margin-top: 1rem; 
+}
 
 .speech-bubble {
     opacity: 0;
@@ -24,10 +24,7 @@ $light-gold: #E2CA78;
     border-radius: 0.5rem; 
     box-shadow: 1rem 1rem rgba($white, 0.08); 
     padding: 1rem; 
-    display:flex;
-    align-items: center; 
-    justify-content: center;
-    text-align: center; 
+    margin-top: -2rem;  
     font-size: 1.5rem; 
     position: relative; 
 
@@ -38,6 +35,8 @@ $light-gold: #E2CA78;
         height: 0.75rem;
         border-radius: 1rem;  
         background-color: $gold;
+        top: calc(100% + 1rem);
+        left: 0%;
         box-shadow: 1.2rem 1.2rem rgba($white, 0.08); 
         animation: hover-3 5s linear infinite; 
 
@@ -49,150 +48,19 @@ $light-gold: #E2CA78;
         height: 0.75rem;
         border-radius: 100%;  
         background-color: $gold; 
+        top: calc(100% + 2.5rem);
+        left: 0%;
         box-shadow: 1.2rem 1.2rem rgba($white, 0.08); 
         animation: hover-4 5s linear infinite; 
     }
 }
 
-.first-bubble {
-    height: 10rem; 
-    width: 12rem; 
-    margin-top: -2rem; 
-    margin-left: 0rem; 
-    
-    &::before {
-        top: calc(100% + 1rem);
-        left: 0%;
-    }
-    &::after {
-        top: calc(100% + 2.5rem);
-        left: 0%;
-    }
-}
-
-.first-bubble-fade-in {
+.speech-bubble-fade-in {
     animation: hover-2 5s linear infinite, fade-in 1s linear forwards; 
 }
 
-.first-bubble-fade-out {
+.speech-bubble-fade-out {
     animation: hover-2 5s linear infinite, fade-out 1s linear forwards; 
-}
-
-.second-bubble {
-    height: 10rem; 
-    width: 16rem;
-    padding: 1.5rem !important;  
-    margin-left: 2rem; 
-    margin-top: 12rem; 
-    display: flex; 
-    flex-direction: column;
-    
-    &::before {
-        top: calc(0% - 2rem);
-        left: 0%;
-    }
-    &::after {
-        top: calc(0% - 3.5rem);
-        left: 0%; 
-    }
-}
-
-.second-bubble-fade-in {
-    animation: hover-2 5s linear infinite, fade-in 1s 1s linear forwards; 
-}
-
-.second-bubble-fade-out {
-    animation: hover-2 5s linear infinite, fade-out 1s linear forwards; 
-}
-
-.info-bubble {
-    // min-width: 30rem;
-    width: 30rem; 
-    min-height: 15rem; 
-    top: 10rem;
-    display: flex;
-    flex-direction: column; 
-    &::before {
-        top: calc(60% - 4rem);
-        left: -1.5rem;
-        height: 4rem !important; 
-        width: 0.75rem !important;
-    }
-    &::after {
-        top: calc(60% - 2rem);
-        left: -2.5rem;
-        width: 0.75rem !important; 
-        border-radius: 1rem !important; 
-        height: 1.5rem !important; 
-    }
-
-    & h2 {
-        margin-top: 4rem;
-        margin-bottom: 0.5rem;  
-    }
-
-    & p {
-        // text-align: left; 
-        font-size: 1.25rem; 
-        background-color: rgba($light-gold, 1);
-        border-radius: 0.5rem; 
-        padding: 1rem; 
-        box-shadow: inset 0 0 0.25rem 0.25rem rgba($black, 0.2); 
-    }
-}
-
-.info-bubble-fade-in {
-    animation: hover-2 5s linear infinite, fade-in 1s 1.1s linear forwards; 
-}
-
-.info-bubble-fade-out {
-    animation: hover-2 5s linear infinite, fade-out 1s linear forwards; 
-}
-
-.features-list {
-    list-style: none;
-    display: flex;
-    flex-direction: row;
-    padding: 0;
-    margin: 0.5rem 0 0 0; 
-}
-
-.feature-btn {
-    color: $black;
-    background-color: $gold; 
-    border: 1px solid $black;
-    border-radius: 0.5rem; 
-    width: 3rem;
-    height: 3rem; 
-    margin-right: 0.5rem;
-    margin-left: 0.5rem; 
-    transition: background-color 0.2s linear, color 0.2s linear, border 0.2s linear; 
-
-    &:hover {
-        background-color: $light-gold; 
-        color: rgba($black, 0.8);
-        border: rgba($black, 0.8); 
-    }
-}
-
-.back-btn {
-    border: none;
-    background-color: $gold;
-    font-size: 1.25rem;
-    position: absolute;
-    border-radius: 0.5rem; 
-    padding: 0;
-    margin: 0; 
-    top: 1rem;
-    left: 1rem; 
-    &:hover {
-        color: $white; 
-    }
-}
-
-#second-bubble {
-    margin-top: 10rem; 
-    margin-left: 10rem; 
 }
 
 @keyframes fade-in {
@@ -259,77 +127,39 @@ $light-gold: #E2CA78;
     }
 }
 
-@media screen and (min-width: 768.1px) {
-    .home-sprite-wrapper {
-        margin-top: 15rem; 
+@media screen and (max-width: 1024px) {
+    .optional-margin {
+        margin-top: 4rem; 
     }
-    .first-bubble {
-        margin-top: 0rem; 
-    }
-    .second-bubble {
-        margin-top: 15rem;
+    .speech-bubble {
+        font-size: 1.2rem; 
+        &::before {
+            height: 0.5rem; 
+            width: 3rem; 
+        }
+        &::after {
+            height: 0.5rem;
+            width: 0.5rem; 
+        }
     }
 }
 
 @media screen and (max-width: 768px) {
-    .sprite-row {
-        margin-top: 10rem; 
-    }
 
-    .home-sprite-wrapper {
-        width: 10rem; 
-    }
-
-    .speech-bubble {
-        font-size: 1rem; 
-    }
-
-    .first-bubble {
-        width: 7rem;
-        height: 8rem; 
-    }
-
-    .second-bubble {
-        width: 7rem;
-        height: 8rem; 
-    }
-
-    .info-bubble {
-        top: 0;
-        width: 20rem;
-        height: 23rem; 
-
-        & p {
-            font-size: 1rem; 
-        }
-    }
-
-    .feature-btn {
-        width: 2rem;
-        height: 2rem; 
-    }
 }
 
 @media screen and (max-width: 500px) {
-
-    .sprite-row {
-        margin-top: 5rem !important; 
-    }
-    
-    .home-sprite-wrapper {
-        width: 6.5rem; 
-    }
     .speech-bubble {
-        font-size: 1rem; 
+        font-size: 1.2rem; 
+        &::before {
+            display: none; 
+        }
+        &::after{
+           display: none; 
+        }
     }
-    .first-bubble {
-        margin-left: -1.5rem; 
-        width: 6rem;
-        height: 6rem; 
-    }
-    .second-bubble {
-        width: 6rem;
-        height: 7rem; 
-        margin-left: -6rem; 
+
+    .optional-margin {
+        margin-top: 4rem; 
     }
 }
